Restores Failures with SAS
Because the Get-DbaCredential can never find the credential, the restores will always fail.  The problem is that Get-DbaCredential is using $matches[0].trim('/') for the name, but that variable is never created.  It appears as if the lines above were supposed to create the variable, but were changed to '$null' at some point.